Codestyle

+++ src/table/table.cpp
@@ -25,7 +25,7 @@ Table::Table(TableBuilder builder) {
 std::string Table::get_sql() {
     std::string answer;
     answer = "CREATE TABLE \'" + name_ + "\' (";
-    for (int i = 0; i < columns_.size(); i++){
+    for (int i = 0; i < (int) columns_.size(); i++){
         if (i > 0){
             answer += ", ";
         }
@@ -38,7 +38,7 @@ std::string Table::get_sql() {
 std::string Table::Save() {
     std::string table_string;
     table_string = name_ + " " + std::to_string(columns_.size()) + " ";
-    for (int i = 0; i < columns_.size(); i++){
+    for (int i = 0; i < (int) columns_.size(); i++){
         table_string += columns_[i]->save();
     }
     return table_string;
